requirements.txt
binance-connector==3.0.0
pytest==7.3.1
python-dotenv==1.0.0
